configs iniciales
