+++ lifecheck-front/drizzle.config.ts
@@ -1,0 +1,5 @@
+import type { Config } from "drizzle-kit";
+export default {
+  schema: "./src/schema.ts",
+  out: "./drizzle",
+} satisfies Config;
