+++ lifecheck-front/package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "generate": "drizzle-kit generate:pg"
   },
   "dependencies": {
     "@auth0/nextjs-auth0": "^3.5.0",
@@ -19,9 +20,11 @@
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
+    "drizzle-orm": "^0.30.9",
     "framer-motion": "^11.1.5",
     "lucide-react": "^0.370.0",
     "next": "14.1.4",
+    "postgres": "^3.4.4",
     "react": "^18",
     "react-dom": "^18",
     "react-speech-recognition": "^3.10.0",
@@ -32,6 +35,7 @@
     "@types/node": "^20.12.7",
     "@types/react": "18.2.79",
     "autoprefixer": "^10.0.1",
+    "drizzle-kit": "^0.20.17",
     "eslint": "^8",
     "eslint-config-next": "14.1.4",
     "postcss": "^8",
